Refactor out base hand definition
